feature: added pagination to ItemAdapter.kt -> ItemPagingAdapter.kt * deps: Jetpack Paging 3 * added: PagingSource class to get the data from api by page

+++ app/src/main/java/com/lefarmico/moviesfinder/data/manager/useCase/MovieBriefListApiUseCase.kt
@@ -4,10 +4,10 @@ import com.lefarmico.moviesfinder.data.entity.MenuItem
 import com.lefarmico.moviesfinder.data.http.response.entity.State
 import kotlinx.coroutines.flow.Flow
 
-fun interface GetPopularMovieBriefList : suspend (Int) -> Flow<State<MenuItem.Movies>>
+fun interface GetPopularMovieBriefList : suspend () -> Flow<State<MenuItem.Movies>>
 
-fun interface GetUpcomingMovieBriefList : suspend (Int) -> Flow<State<MenuItem.Movies>>
+fun interface GetUpcomingMovieBriefList : suspend () -> Flow<State<MenuItem.Movies>>
 
-fun interface GetTopRatedMovieBriefList : suspend (Int) -> Flow<State<MenuItem.Movies>>
+fun interface GetTopRatedMovieBriefList : suspend () -> Flow<State<MenuItem.Movies>>
 
-fun interface GetNowPlayingMovieBriefList : suspend (Int) -> Flow<State<MenuItem.Movies>>
+fun interface GetNowPlayingMovieBriefList : suspend () -> Flow<State<MenuItem.Movies>>
